+++ src/pantainos/utils/__init__.py
@@ -2,8 +2,10 @@
 Utility modules for Pantainos
 """
 
+from .runner import ApplicationRunner
 from .testing import create_mock_event
 
 __all__ = [
+    "ApplicationRunner",
     "create_mock_event",
 ]